chore: deprecate useSurveyList hhook, merge it into useSurvey. wrap SurveyProvider in App's root
