+++ apps/frontend/src/pages/SurveyStudio.tsx
@@ -1,10 +1,5 @@
-import { SurveyProvider } from "@/modules/surveys/SurveyProvider";
 import Studio from "@/modules/surveys/components/Studio";
 
 export function SurveyStudio() {
-    return (
-        <SurveyProvider>
-            <Studio />
-        </SurveyProvider>
-    );
+    return <Studio />;
 }
